[WIP]update: move analysis part to engine
