+++ findit/core.py
@@ -1,11 +1,11 @@
-import cv2
 import os
 import numpy as np
 import typing
 import json
 
 from findit.logger import logger, LOGGER_FLAG
-import findit.toolbox as toolbox
+from findit import toolbox
+from findit.engine import TemplateEngine, FeatureEngine
 
 
 class FindIt(object):
@@ -82,17 +82,12 @@ class FindIt(object):
     def find(self,
              target_pic_path: str = None,
              target_pic_object: np.ndarray = None,
-             mask_pic_path: str = None,
-             mask_pic_object: np.ndarray = None,
-             scale: typing.Sequence = None):
+             *args, **kwargs):
         """
         start match
 
         :param target_pic_path: '/path/to/your/target.png'
         :param target_pic_object: your_pic_cv_object (loaded by cv2)
-        :param mask_pic_path: '/path/to/your/target.png'
-        :param mask_pic_object: your_pic_cv_object (loaded by cv2)
-        :param scale: default to (1, 3, 10)
         :return:
         """
 
@@ -104,43 +99,22 @@ class FindIt(object):
         logger.info('start finding ...')
         target_pic_object = toolbox.pre_pic(target_pic_path, target_pic_object)
 
-        # mask
-        if mask_pic_path or mask_pic_object is not None:
-            logger.info('mask detected')
-            mask_pic_object = toolbox.pre_pic(mask_pic_path, mask_pic_object)
-
-        # scale
-        if not scale:
-            # default scale
-            scale = (1, 3, 10)
-        logger.info('scale: {}'.format(str(scale)))
+        # engine init
+        template_engine = TemplateEngine(*args, **kwargs)
+        feature_engine = FeatureEngine(*args, **kwargs)
 
         result: typing.List[dict] = list()
         for each_template_path, each_template_object in self.template.items():
             logger.debug('start analysing: [{}] ...'.format(each_template_path))
 
-            # template matching
-            min_val, max_val, min_loc, max_loc = self._compare_template(
-                target_pic_object,
-                each_template_object,
-                scale,
-                mask_pic_object
-            )
-            logger.debug('raw compare result: {}, {}, {}, {}'.format(min_val, max_val, min_loc, max_loc))
-            min_loc, max_loc = map(lambda i: toolbox.fix_location(each_template_object, i), [min_loc, max_loc])
-            logger.debug('fixed compare result: {}, {}, {}, {}'.format(min_val, max_val, min_loc, max_loc))
-
-            # feature matching
-            feature_center_point = self._compare_feature(target_pic_object, each_template_object)
+            template_match_result = template_engine.execute(each_template_object, target_pic_object, *args, **kwargs)
+            feature_match_result = feature_engine.execute(target_pic_object, each_template_object)
 
             # add to result list
             current_result = {
                 'path': each_template_path,
-                'min_val': min_val,
-                'max_val': max_val,
-                'min_loc': min_loc,
-                'max_loc': max_loc,
-                'feature_loc': feature_center_point,
+                'template': template_match_result,
+                'feature': feature_match_result,
             }
             logger.debug('result for [{}]: {}'.format(each_template_path, json.dumps(current_result)))
             result.append(current_result)
@@ -152,85 +126,6 @@ class FindIt(object):
         }
         logger.info('result: {}'.format(json.dumps(final_result)))
         return final_result
-
-    def _compare_template(self,
-                          target_pic_object: np.ndarray,
-                          template_pic_object: np.ndarray,
-                          scale: typing.Sequence,
-                          mask_pic_object: np.ndarray = None) -> typing.Sequence[float]:
-        """
-        compare via template matching
-        (https://www.pyimagesearch.com/2015/01/26/multi-scale-template-matching-using-python-opencv/)
-
-        :param target_pic_object:
-        :param template_pic_object:
-        :param scale: default to (1, 3, 10)
-        :param mask_pic_object:
-        :return: min_val, max_val, min_loc, max_loc
-        """
-        pic_height, pic_width = target_pic_object.shape[:2]
-        result_list = list()
-
-        for each_scale in np.linspace(*scale):
-            # resize template
-            resize_template_pic_object = toolbox.resize_pic_scale(template_pic_object, each_scale)
-
-            # resize mask
-            if mask_pic_object is not None:
-                resize_mask_pic_object = toolbox.resize_pic_scale(mask_pic_object, each_scale)
-            else:
-                resize_mask_pic_object = None
-
-            # if template's size is larger than raw picture, break
-            if resize_template_pic_object.shape[0] > pic_height or resize_template_pic_object.shape[1] > pic_width:
-                break
-
-            res = cv2.matchTemplate(
-                target_pic_object,
-                resize_template_pic_object,
-                self._cv_method_code,
-                mask=resize_mask_pic_object)
-            result_list.append(cv2.minMaxLoc(res))
-
-        logger.debug('scale search result: {}'.format(result_list))
-        # return the max one
-        return sorted(result_list, key=lambda i: i[1])[-1]
-
-    def _compare_feature(self,
-                         target_pic_object: np.ndarray,
-                         template_pic_object: np.ndarray) -> typing.Sequence[float]:
-        """
-        compare via feature matching
-
-        :param target_pic_object:
-        :param template_pic_object:
-        :return:
-        """
-        # Initiate SIFT detector
-        sift = cv2.xfeatures2d.SIFT_create()
-
-        # find the keypoints and descriptors with SIFT
-        _, des1 = sift.detectAndCompute(template_pic_object, None)
-        kp2, des2 = sift.detectAndCompute(target_pic_object, None)
-
-        # BFMatcher with default params
-        bf = cv2.BFMatcher()
-        matches = bf.knnMatch(des1, des2, k=2)
-
-        # Apply ratio test
-        good = []
-        for m, n in matches:
-            if m.distance < 0.75 * n.distance:
-                good.append([m])
-        point_list = list()
-        for each in good:
-            img2_idx = each[0].trainIdx
-            point_list.append(kp2[img2_idx].pt)
-
-        # cal the central point
-        center_x = sum([_[0] for _ in point_list]) / len(point_list)
-        center_y = sum([_[1] for _ in point_list]) / len(point_list)
-        return center_x, center_y
 
     def reset(self):
         """ reset template, target and result """
